requirements.txt
# Core Django
Django==6.0
asgiref==3.11.0
sqlparse==0.5.5

# Django Channels for real-time WebSocket messaging
channels==4.3.2
daphne==4.2.1
Twisted==25.5.0

# Production Server & Static Files
gunicorn==23.0.0
whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10

# AI Chat Support
groq==1.0.0
google-generativeai==0.8.0

# Templates
Jinja2==3.1.6

# CORS
django-cors-headers==4.9.0

# Storage (S3/Backblaze B2)
django-storages==1.14.6
boto3

# Environment variables
python-dotenv==1.0.1
python-decouple==3.8

# Stock data - Yahoo Finance
yfinance
pandas
numpy
openpyxl  # For Excel file (.xlsx) support

# Django Import-Export for admin data import/export
django-import-export==4.3.1

# HTTP requests
requests
beautifulsoup4

# Date/Time handling
python-dateutil
pytz

# WebSockets
websockets

# Security
pyOpenSSL
cryptography
service-identity

# Utilities (let pip resolve versions)
attrs
packaging
six
typing_extensions
pydantic

protobuf>=3.20.2,<6.0.0
polib==1.2.0

# Django Allauth - Authentication & Social Login
django-allauth

# HTMX Integration
django-htmx==1.21.0

# OTP and SMS
pyotp
twilio
django-otp
qrcode
